requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-localflavor==2.1
django-rest-auth==0.9.3
djangorestframework==3.9.1
idna==2.8
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
